Add countries to response dto. Moved SmartDoubleSerializer to util folder. Refactor logic of fetchPhotos method
