+++ src/main/java/org/cyberrealm/tech/muvio/repository/MediaRepository.java
@@ -49,6 +49,6 @@ public interface MediaRepository extends MongoRepository<Media, String>, MediaRe
     Slice<MediaBaseDto> findByTitle(String title, Pageable pageable);
 
     @Query(value = "{}", fields = "{ 'id': 1, 'title': 1, 'genres': 1, 'rating': 1, "
-            + "'posterPath': 1, 'duration': 1, 'releaseYear': 1  'type': 1 }")
+            + "'posterPath': 1, 'duration': 1, 'releaseYear': 1, 'type': 1 }")
     List<MediaBaseDto> getAll(Pageable pageable);
 }
